fix: convert Integer to Long

See merge request jack-hagedorn/legalt-backend!12

+++ src/main/java/jact/lagaltproject/exceptions/ProjectNotFoundException.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(value = HttpStatus.NOT_FOUND)
 public class ProjectNotFoundException extends RuntimeException{
 
-    public ProjectNotFoundException(int id) {
+    public ProjectNotFoundException(Long id) {
         super("Project does not exist with that ID: " + id);
     }
